gen-ai/orchestrator-server/src/main/python/tock-llm-indexing-tools/pyproject.toml
[tool.poetry]
name = "tock-llm-indexing-tools"
version = "0.1.0"
description = "Tools to index data from different sources to TOCK RAG vector DB"
authors = ["Pierre Guirriec <pierre.guirriec@partnre.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
docopt-ng = "^0.9.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
beautifulsoup4 = "^4.12.2"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
opensearch-py = "^2.7.1"
gen_ai_orchestrator = {path = "../server", develop=true}
langchain-community = "^0.3.7"
pre-commit = "^4.0.1"
aiometer = "^0.5.0"
aiohttp = "^3.11.6"
aiohttp-socks = "^0.10.1"
pydantic = "^2.9.2"
humanize = "^4.11.0"
langfuse = "^2.57.7"
ragas = "^0.2.13"
rapidfuzz = "^3.12.2"
sacrebleu = "^2.5.1"
rouge_score = "^0.1.2"
python-pptx = "^1.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
